Update `svb_client` to `svb_api_client` to be more specific in import.
