+++ lib/api/account_balance.rb
@@ -1,13 +1,11 @@
-module SVB
-  module AccountBalance
+module AccountBalance
 
-    ACCOUNT_BALANCE_SCOPE = 'accountbalance'
+  ACCOUNT_BALANCE_SCOPE = 'accountbalance'
 
-    def get_account_balance(account_number:)
-      raise ArgumentError, "Account number is required" if account_number.blank?
+  def get_account_balance(account_number:)
+    raise ArgumentError, "Account number is required" if account_number.blank?
 
-      body = {account: account_number}.to_json
-      post(path:"/v1/accounts/balances", body: body, scope: ACCOUNT_BALANCE_SCOPE)
-    end
+    body = {account: account_number}.to_json
+    post(path:"/v1/accounts/balances", body: body, scope: ACCOUNT_BALANCE_SCOPE)
   end
 end
