--- svbclient.gemspec
@@ -1,13 +1,0 @@
-Gem::Specification.new do |s|
-  s.name        = 'svbclient'
-  s.version     = '4.0.0'
-  s.date        = '2024-10-01'
-  s.summary     = "Silicon Valley Bank Ruby SDK"
-  s.description = "Silicon Valley Bank Ruby SDK"
-  s.authors     = ["Streamlined Financial Inc."]
-  s.email       = 'apisupport@svb.com'
-  s.files       = ["lib/svb_client.rb"]
-  s.homepage    = 'http://rubygems.org/gems/svbclient'
-  s.license     = 'MIT'
-  s.add_dependency 'rest-client'
-end
